requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
supabase==2.10.0
asyncpg==0.30.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.0
solana==0.35.0
solders==0.22.0
celery==5.4.0
redis==5.2.0
apscheduler==3.10.4
python-multipart==0.0.17
orjson==3.10.11
structlog==24.4.0
sentry-sdk[fastapi]==2.19.0
